pyproject.toml
[tool.poetry]
name = "streamlit-case"
version = "0.1.0"
description = "Projeto para visualização de dados com Streamlit e Supabase"
authors = ["João Escórcio <jv.escorcio@gmail.com>"]
license = "MIT"
package-mode = false

# Definindo a versão compatível do Python
[tool.poetry.dependencies]
python = ">=3.9,<4.0"
pandas = "^2.2.3"
supabase = "^2.15.0"
python-dotenv = "^1.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
run = "script:main"

[tool.poetry.group.dev.dependencies]
pytest = "^7.3.0"
